+++ CHANGELOG.md
@@ -8,6 +8,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Update
 - Fixed Issue with page sizing with empty results
 
+## [2.4.9] - 2025-05-14
+### Added
+- Added UPC mapping functionality to order exports
+- Added ability to select product attribute to use as UPC value in order exports
+
 ## [2.5.1-beta] - 2024-11-01
 ### Added
 - Updated custom store export implementation to include the payment method.